+++ flag/service/prometheus/prometheus.go
@@ -6,6 +6,7 @@ type Prometheus struct {
 	BaseDomain              string
 	Bastions                string
 	EvaluationInterval      string
+	ImageRepository         string
 	LogLevel                string
 	Mayu                    string
 	Retention               PrometheusRetention
@@ -15,5 +16,4 @@ type Prometheus struct {
 
 type PrometheusRetention struct {
 	Duration string
-	Size     string
 }